requirements.txt
streamlit==1.41.1
pandas==2.2.3
psycopg2-binary==2.9.9